+++ java-client/src/main/java/co/elastic/clients/elasticsearch/core/IndexRequest.java
@@ -41,7 +41,6 @@ import co.elastic.clients.util.ObjectBuilder;
 import jakarta.json.JsonValue;
 import jakarta.json.stream.JsonGenerator;
 import java.lang.Boolean;
-import java.lang.Integer;
 import java.lang.Long;
 import java.lang.String;
 import java.util.HashMap;
@@ -65,7 +64,7 @@ public final class IndexRequest<TDocument> extends RequestBase implements JsonpS
 	private final Long ifPrimaryTerm;
 
 	@Nullable
-	private final Integer ifSeqNo;
+	private final Long ifSeqNo;
 
 	@Nullable
 	private final OpType opType;
@@ -173,7 +172,7 @@ public final class IndexRequest<TDocument> extends RequestBase implements JsonpS
 	 * API name: {@code if_seq_no}
 	 */
 	@Nullable
-	public Integer ifSeqNo() {
+	public Long ifSeqNo() {
 		return this.ifSeqNo;
 	}
 
@@ -311,7 +310,7 @@ public final class IndexRequest<TDocument> extends RequestBase implements JsonpS
 		private Long ifPrimaryTerm;
 
 		@Nullable
-		private Integer ifSeqNo;
+		private Long ifSeqNo;
 
 		@Nullable
 		private OpType opType;
@@ -392,7 +391,7 @@ public final class IndexRequest<TDocument> extends RequestBase implements JsonpS
 		 * <p>
 		 * API name: {@code if_seq_no}
 		 */
-		public Builder<TDocument> ifSeqNo(@Nullable Integer value) {
+		public Builder<TDocument> ifSeqNo(@Nullable Long value) {
 			this.ifSeqNo = value;
 			return this;
 		}
